Move deprecated `Stdin` and `Stdout` classes to internal `Io` namespace
